+++ ephem/cities.py
@@ -150,27 +150,15 @@ def city(name):
     return o
 
 def lookup(address):
-    """Given a string `address`, do a Google lookup and return an Observer.
+    """DEPRECATED: Google, alas, no longer supports anonymous lat/lon lookup.
 
-    Avoid calling this very often, to honor Google's terms of service.
-    Instead you can run it once, print out the result, and cut and paste
-    the Observer back into your code to use as often as you like!
+    Because looking up an address is really a problem in geography, not
+    astronomy, PyEphem is not planning on repairing this routine.  Look
+    for a good Python geolocation library if you need to turn strings
+    into latitudes and longitudes.
 
     """
-    parameters = urlencode({'address': address, 'sensor': 'false'})
-    url = 'http://maps.googleapis.com/maps/api/geocode/json?' + parameters
-    data = json.loads(urlopen(url).read().decode('utf-8'))
-    results = data['results']
-    if not results:
-        raise ValueError('Google cannot find a place named %r' % address)
-    address_components = results[0]['address_components']
-    location = results[0]['geometry']['location']
-
-    o = ephem.Observer()
-    o.name = ', '.join(c['long_name'] for c in address_components)
-    o.lat = radians(location['lat'])
-    o.lon = radians(location['lng'])
-    return o
+    raise NotImplementedError(lookup.__doc__.split(None, 1)[1].strip())
 
 def lookup_with_geonames(q, username):
     """Given a string `q`, do a geonames lookup and return an Observer.